requirements.txt
decouple==0.0.7
Django==4.0
django-filter==22.1
django-import-export==3.0.1
django-restframework==0.0.1
django-userforeignkey==0.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
python-decouple==3.6
Pillow==9.2.0
requests==2.28.1
whitenoise==6.2.0

